[DICE profile I-D] Brian Carpenter's Gen-ART review comment
Gen-ART Last Call review of draft-ietf-dice-profile-14:

Oh. I find the registry very confusing. It would help ignorant people like
me if the sentence read

"The IPv6 multicast address used for CoAP site-local discovery is FF02::FD."

except that there *is* a bug in that sentence: ff02:: is the link-local range,
site-local would be ff05::fd. So it should be

"The IPv6 multicast address used for CoAP link-local discovery is FF02::FD."